Add tcpbridge
tcpbridge is a little Python script to connect together two TCP sockets
with each other. The idea is to use qemus socket devices to be able to
bridge together NICs. Let's say we have two qemu instances running and
both are using TCP sockets, so they are listening to port 10001. Now we
can use tcpbridge to connect to both qemu and whenever we receive some
traffic we send it over to the other qemu instance, thus bridging the
two NICs together.

By letting all qemu instances only listen for TCP we can connect,
disconnect and modify the connections afterwards simply by restarting
tcpbridge with a different configuration. If we were to use the UDP mode
of qemu or TCP client then we would have to know the topology before we
start the instances.