Add arg parsing to tcpbridge
Instead of taking a config file we work with command line argumens. Currently
supporting a --p2p option which creates a point-to-point bridge between two
virtual routers. The format is --p2p X/Y-Z/N where X is the id of the first
router, Y is the interface number on that router. Z is the id of the second
router and N is the interface number on the second router. For example, if
we've started router 10 and 11 and wish to connect them together on their first
interface we would use: --p2p 10/1-11/1

0 is actually the lowest interface id but it's used for management.