Replace mgmt bridge with user networking
Instead of using a Linux bridge for the mgmt interface of the router we rely on
qemu user mode networking. When launching the container without --net=host we
can't really use a Linux bridge anyway since it ends up in a different
namespace, although one could potentially use pipework to bridge a Linux bridge
in the host to one inside the container.

qemu user mode networking is a bit special and uses NAT per default to allow
outbound connections from the VM to the rest of the world. We are mostly
interested in the other direction so we can reach port 22 and 830 of the
virtual router. Due to some quirk that I have been unable to identify I cannot
simply map host 22 to VM:22 because then exposing port 22 in docker leads to
hung TCP connections. Instead I map 2022->VM:22 and have a socat process that
maps host:22 to host:2022.. which works. Super weird and useless use of (so)cat
but it works.