Switch to random MAC addresses
numeric-id was used to compute a MAC address but since we want to remove
numeric-id we switched over to simply generating semi-random MAC addresses.
They are all in the OUI space of qemu (52:54:00) and the last octet is the same
as the port so you have a chance of knowing wether you are connected to the
correct router.

Part of #2.