Dynamic number of interfaces on vMX
In earlier JUNOS versions the number of NICs is limited to 23. As I have
another system with an old YANG model I'm currently lowering the number of NICs
to 20 to be compatible with that system. Once it has a newer YANG model we
should be able to raise the limit.

This doesn't really affect anything negatively since the value of 96 is only
used to tell JUNOS how many NICs it should have, the number of NICs provisioned
by KVM has always been 20 so we never had more than 20 usable NICs. To actually
go beyond 20 I believe we need to add more PCI busses and stuff.

Part of #31.