xrv: simplify bootstrap process
Cleanup of xrv launch script by removing unused functionality. Like we don't
need to tickle the device because we know we are connected from the second the
router is started (since we also started it). Nor do we need to do non-blocking
start since the script is the only thing running.