vmx: refactor & simplify for more resilient start
Cleaned up some stuff that's not really needed in a vrnetlab container (legacy
from old netlab environment) making the script simpler.

The start functions have been split up and we now carry a telnet connection to
both the VCP and vFPC as well as references to the processes which should make
it easier to have more control over them. We use this control to implement
separate startup monitoring of the vFPC and we should hopefully be able to
restart it, should it fail.

Fixes #44.