vmx: Delay start of vFPC
The vFPC starts rather quickly, typically under half a minute on my laptop so
we can delay starting it until the control plane is ready or near ready.
Currently the only time we can determine 'near ready' is when we just received
the login prompt from the VCP. It probably takes a little bit longer to start
it so we won't delay the complete startup time by much.

The advantage of delaying the vFPC start is that the vFPC consumes large
amounts of CPU since it's busy-looping waiting for packets. It's assigned three
CPU cores but typically uses 100-200% CPU. In low load scenarios, like starting
a single vMX on a four core machine (like my laptop) the startup time is
virtually identical, although somewhat surprisingly I've actually seen a very
slight decrease with this patch. In high load scenarios however, where many
virtual routers compete for fewer resources, the difference seems to be around
25-30% faster boot.