topom: make --run more deterministic
At least the order of the `docker run` commands is now ordered and is
deterministic. The tcpbridge run command will differ since it iterates over
links whose order is determined by looping over dicts, i.e. it is not
deterministic. The code in run_topology() itself though should now be
deterministic so as long as we fix other places we should get deterministic
output.

Part of #51.