vmx: fix support for vMX version 16
vMX version 15 have use the three first interfaces on the vFPC for "special
purposes", like first is mgmt, the second is internal communication to the
control plane and the third.. well, I'm not sure what it's used for but we
create a dummy interface for it so that the first real traffic port becomes the
fourth interface and this maps to ge-0/0/0. vMX version 16 on the other hand
only has the management port and the internal communication making the third
interface map to ge-0/0/0.

This means we need to insert a dummy interface only for version 15 and so I've
refactored a few things so that we may extract the version from the VCP image
filename.

Fixes #55.