vmx: Use lite image when available
Older versions of vMX ship with two vFPC images - one is for performance and
the other is the lite image meant for low throughput labing. The performance
image requires some specific CPU instructions and I couldn't actually get it to
work on my dev laptop, despite it being rather new, so switching to the lite
image.

vMX 14.1 is the only versions I've seen that actually has these two images.

Fixes #30.