tcpbridge: new mode to connect tcp to a tap intf
The new --vr2tap argument allows the connection of a VR router, with its
interfaces exposed as TCP sockets, to a a tap interface so that it's possible
to easily communicate with a router from a user space process.

The easiest way to use vr2tap mode is to put tcpbridge into a docker image and
upon start run tcpbridge in vr2tap mode. A tap0 interface will be created and
the routing inside the container. Add an address and you can now communicate
with the router!

Fixes #58.