vr-bgp: add received route API method
It's now possible to poll vr-bgp of the currently announced routes. A second
process is used to receive the route from ExaBGP and this information is then
written to a sqlite database. The bgpapi process has a new /received end point
which reads the sqlite database and returns the received routes.

Fugly code and not much error checking but hey it works ;)