vr-xcon: listen - don't connect for tap mode
The old --vr2tap mode would actively connect to a remote endpoint but
that kind of breaks the philosophy we have of that the we shouldn't
require the topology to be known on startup. --vr2tap has thus been
replaced with --tap-listen which works the same way in that it bridges
packets to a tap interface but it sits passively listening to an
incoming connection. vr-xcon --p2p can be used to connect one of these
listening sockets to a virtual router.

Fixes #62.