sros: fix KVM error detection
I implemented the check_qemu function awhile to detect KVM problems and
restart the VM if any of those arose. Howwever, while it does work I
noticed it does trigger on the correct event. Instead of seeing the KVM
error it will restart the VM when there haven't been any output for some
time.

This change fixes so that we capture STDERR and any output on STDERR
means we restart the VM - hope it's not too aggresive but I don't think
any normal output comes from STDERR.