xrv: completely refactor startup script
Move out the generic parts of the XRV startup script into the common
vrnetlab module. Also introduces the VM class, which is intended to
model a VM. Some virtual routers use multiple VMs so it's useful to be
able to distinguish the multiple VMs from the single VR.

Part of #28.
Part of #59.