XRv: use 128 NICs per default
With the changes to the common startup script it is now possible to use many
NICs. I've done very limited testing of XRv with 128 NICs but it boots up and I
have basic connectivity without the sudden death issue we have seen at times
with too many NICs on one PCI bus.

Fixes #32.