Do vMX VCP install in docker build
When the vMX VCP boots for the first time it will read a config file that
instructs the VCP in which mode it should run, for example VRR or be the VCP in
a vMX. After this config file is read it will restart. This takes a long time
and since we put vMX in docker it will look like every boot is the first boot,
thus always taking a long time.

This change moves the first boot of the VCP from run time to build time. It
will only start the VCP (not vFPC), let it have its initial config, restart and
then shut it down. docker run of the image will be much much faster.

--privileged isn't available in docker build which means the VCP have to run
without hardware acceleration. This in turn means that the build process is
excruciatingly slow but I think it's worth it compared to always having slow
startup of the virtual router.

Fixes #41