Don't regenerate crypto keys on XR
If we have already generated crypto keys we don't have to do it again. For
example, if we restart the docker container the bootstrap script will be run
again and while this means the exact same configuration will be applied
(~idempotent) we would halt on the crypt key generation since we were prompted
by the router to overwrite the last ones. Now we intelligently answer 'no' to
this :)

Fixes #87.