Add debug output for XRv
I'm seeing some strange failures where the bootstrap configuration fails on XRv
routers, in particular the configuration of the management interface fails and
I suspect we are so quick in trying to configure it the router hasn't properly
started yet (despite saying so). I'm adding this `show interface description`
so that we can get see in the trace log if this is indeed the case.