XRv: retry entering config-if sub-mode
XRv sometimes fails (hunch is like 1 of 20 in our CI system) to enter the
config-if sub mode for the MgmtEth interface. I suppose it's some form of race,
e.g. we try to configure the interface before the router is really ready. I've
added a wrapper around this so that we can try to catch the error and retry.