Speed up vMX installation
The vMX installation is a little complex as we actually fire up the RE VM
during the installation process itself. As you aren't allowed to use
--privileged for building a docker container it means that the qemu machine is
running without any form of hardware offloading so it's really slow.

By splitting the installation into two steps; the first is just assembling the
basic pieces in a docker image, then running that image with docker run, thus
enabling the use of --privileged and finally committing this final container
instance into the final docker image. The use of --privileged considerably
speeds up the installation.