Separate registry URL check
The Makefiles for vr-xcon is different enough that we can't reuse
makefile.include but we still want reuse of the DOCKER_REGISTRY URL check so I
moved it into a separate file that can then be included.

Also the regex is updated to allow for not just setting the registry FQDN and
port but also include a prefix, like /foo at the end, to match groups in a
GitLab Docker registry.