Avoid static name vr-temp
Instead of letting the container be named vr-temp we let docker use its default
mode of coming up with a name dynamically. To keep track of our container we
instead use the --cidfile option to write the container ID to a file which we
then read to get the id.