Parse next-hop as path attribute
We parse out the next-hop from the information we receive from ExaBGP. My
previous understanding of the information sent from ExaBGP was incorrect and
what I thought was the neighbor address is in fact the next-hop from the NLRI
information.

However, instead of presenting this as NLRI information we just store it as a
(path) attribute in the database and also display it the same way over the API.
This makes it very easy to use and all information is still just keyed on the
prefix. It does however prevent stuff like BGP Add-Path from working with
vr-bgp....