Add --listen argument
We enabled listening by default, which really does make sense. There seems to
be a bug around how we pick up the state of a neighbor however. I'm not
entirely sure if we are doing the wrong thing or if it's a bug in ExaBGP. Until
it has been resolved we are going to disable listening per default and instead
have this --listen option to selectively enable it if and when needed. Once the
neighbor state issue has been resolved I reckon we can listen per default
again.