Use overlay qcow2 file
Docker defaults to using aufs for backing the image layers and while aufs is
copy-on-write (COW) it does that on a per file basis rather than on a byte
level so if you write to a file the entire file will first be copied from the
image layer to the container layer. In the case of the disk image for the
virtual routers this means we will copy that whole thing and since it's often
in the range of hundreds of megabytes or even a gigabyte, that's a lot of data
to copy around for no real good reason. Not to mention the increased disk
usage.

By creating an qcow2 overlay image inside of our container we make sure that
all writes go to the qcow2 overlay and we can avoid that expensive copy.

Fixes #144.