Reduce size of vmx
Previously we copied the image into the docker image then extracted it
which meant that we had one layer with the .tgz file and another with
the extracted content and then removing the .tgz doesn't free any space
so the final images were upwards of 7-8GB.

By moving the extraction to before the build we can avoid this. A little
less elegant to build outside of the docker container but since this is
all really standard tools I don't think it matters much.