Add docker network support.
Routers can have a `docker_network` parameter (for now all have to be in
a single network). All devices will be attached to the docker network
and given a network-alias with the name of the device.

vr-xcon will also be attached to the docker network.

If the network does not exist, it will be created.