Add tcp2raw mode for vr-xcon
This is a copy of the tcp2tap mode but this time using a raw socket
instead of the tap interface so we can easily interface with the veth
interfaces given to us by docker.

Naturally we should consolidate the code between tcp2tap and tcp2raw
since it's largely the same. The same goes for the normal tcpbridge
mode. If we instead built classes that abstracted away the type of
interface it would all just be X to Y, no matter what X or Y is. Anyway,
that's for a day when I don't have deadlines.