Use XRS-40 style
SROS 15 defaults to an XRS-20 chassis but actually assumes a XRS-40
topology, which consists of two XRS-20. It will among other things want
16 SFMs. I had decreased the number of SFMs to 8 to match what we get on
SROS 16 and this seems to have negatively impacted performance of the
router somehow, like doing "show sfm" could take minute to return.

I have now gone the other way and changed so that SROS 16 is run as an
XRS-40!