Fix logic error
Prior to this change, 'make docker-image' was failing for nxos builds because the makefile variable $IMAGE and $VERSION were the same causing the script to fail, but the intent of the conditional that compares these two values appears to have been put in place to fail if they were NOT the same. This change adjusts the conditional to correct this logic error.