Add support for SROS 19 (distributed mode)
The cards we were using in previous versions (`cpm-x20`) are no longer
supported in virtual routers running 19.7.R1. We settled on `xcm-x20`
instead. There are others available, but with this one, interface names
were not changed. This allows us to use the same version of topomachine
regardless of SROS version.

The small tradeoff of using `xcm-x20` is that it requires the router to
run in distributed mode 🤷
Use card=xcm-x20 instead of cpm-x20 for distributed mode for sros > =19.