Disable paging on VRP
We're looking at the running config of the GigabitEthernet interfaces.
If the list grows longer than a single screen, we may miss the interface
we're looking for.

To further complicate matters, VRP seems to "forget" that we set paging
for the currently active session?! To make sure we're not getting paged
output while waiting, send the no_paging_command every time.