Increase RAM on XRV9k & expose as arg
Enabling telemetry on XRV9k we see that the memory use enters a "severe"
state with less than 10% free. At that point, it doesn't accept
configuration changes any more.

It's somewhat weird as xrv9k reports having 4.6GB of RAM in total
whereas we hand the VM 12GB. I think that it uses nested VMs and
internally carves it up into the control plane and forwarding plane, but
who knows? Anyway, increasing the default from 12GB to 16GB now in an
attempt at giving more to the control plane.