Use unique overlay disk image for VM instance
Newer versions of qemu prevent multiple guests directly sharing the same
disk image - a write lock cannot be obtained. While it is possible to
implement proper guest drive sharing using NBD (Network Block Device),
that is likely not something that we should do in the first place. Not
really sure how and why the blind sharing worked in the first place?!