Add boot license and SSH MAC config
For some features, such as crypto, to work we need to configure a
license level prior to booting. Here we set Network Premier with the DNA
Premier add-on. We might not need everything that has to offer, but
since it's honor based, it shouldn't be an issue for now.

Also forcing IP SSH MAC to hmac-sha2-512 since by default IOS-XE seems
to only support some new MAC ciphers that NSO doesn't want to offer.