Ensure we use overlay image as base for overlays
We always use an overlay disk image to avoid the cost of copy-on-writes
when the virtual router persists data. If the container image contains a
pre-installed device then one overlay is already present in the image.
To ensure *it* is used as the base for a *new* overlay image and not
just attached directly, we sort the candidate disk images.

For example, we start with the 'foo.qcow2' initial image. Then install
stage produces 'foo-0-overlay.qcow2' overlay image and both are included
in the pre-installed image. When the vrnetlab container starts the next
time, the launch script picks up 'foo-0-overlay.qcow2' image as the new
base and creates 'foo-0-overlay-0-overlay.qcow2' overlay.