Add new xcon
A littl experimental right now. This xcon is implemented in Acton. I
think the actor model is a great fit here for doing error handling etc.
Quite natural. The whole state tracking was a bit of a POC. Maybe should
be removed.

The new xcon is automatically compiled and included in the container
image so we can easily run it by switching entrypoint!