Prevent socket leak by closing old connection
When we reconnect, we do that by effectively discarding the old
TCPIPConnection actor and replacing it by reassigning our local variable
reference to a new TCPIPConnection actor. The old TCPIPConnection actor
will potentially be garbage collected but as there are is no cleanup
hook, it is unable to clean out the relevant stuff from the I/O
subsystem. By explicitly calling .close() on it first, it should be
cleaned up and we prevent a socket leak.