Use single WriteFile actor for serializing file writes
We have many TcpEndpoint actors all pushing their internal state changes
to the Healthcheck actor with Healthcheck.update(). These calls are
serialized and processed in the order they arrive. The internal
Healthcheck._flush() method then uses a single file.WriteFile actor to
serialize writes to the "/health" file. If we did not do that and used
a *new* file.WriteFile actor directly (like we used to), there is a race
between the different file.WriteFile actors and a chance of inconsistent
state written to the file.