Detect backing image format with qemu-img
Recent versions of QEMU, like the one that ships with Debian Bookworm,
started requiring users to specify the backing image format when
creating an overlay image with 'qemu-img create'. I have seen files
where the extension did not match the format so I guess from the
perspective of QEMU it makes sense to be explicit. To make this
transparent to the users of vrnetlab  we use 'qemu-img info' to detect
the format and feed it back to 'qemu-img create'!?