Extend 'run_args' to support lists, dicts
topomachine already supports 'run_args' for passing a raw string
containing all the arguments to a container. By accepting list or
dictionary values we can keep the key-value pairs structure in JSON and
then pass the arguments as key value pairs to the container. Note how
the key and value are passed verbatim - we make no assumptions about
prefixing the key names with '-' or quoting the values.