Teach topomachine to use static interface assignment
In some rare cases we want the link between devices to be created on a
specific (static) interface. It is now possible to specify the desired
numeric interface ID in the p2p or hub link specifications:

{"p2p": {"pe": ["cpe:3"] }}

Under normal circumstances the "pe" and "cpe" devices would be connected
on the first available interface - 1. We can override "cpe" interface
assignment by providing the value after the name.