Set announced prefix as default route source
Extend the "pingable-address" feature by adding
"pingable-address-as-source": true|false entry for prefixes. If set to
true it updates the default route in the container with
src=pingable-address. This tricks the applications running in the
container to use the loopback address as source for traffic destined to
use the default route.