Add Catalyst 8000v (c8000v) (#181)
* Copy c8000v from vrnetlab upstream

This is a direct copy from the upstream vrnetlab repository and as such
probably does not work with containerlab. Compatibility changes follow
in the next commits.

* Add --hostname argument

* Add --connection-mode argument

* Add STARTUP_CONFIG_FILE

* Fix 'ip domain name' bootstrap config on csr1000v / cat8000v

The 'ip domain-name FOO' command was renamed to 'ip domain name FOO' in
newer versions. We pass the VERSION makefile variable to the vrnetlab
launcher to keep the version extraction logic in one place.

* Align c8000v Dockerfile with others

Use ubuntu:20.04 base and install other common tools.

* Remove unused imports

* Open file "with" statement

* formatted with ruff and switch to debian bookworm for base image

also use md syntax

* squash some pylint warning about the log and var interpolation

---------

Co-authored-by: Roman Dodin <dodin.roman@gmail.com>